+++ react-design/src/hooks/useTaskList.ts
@@ -1,0 +1,42 @@
+import { useState, useEffect } from "react";
+import type { Task } from "../domain/Task";
+
+export function useTaskList() {
+  const [tasks, setTasks] = useState<Task[]>(() => {
+    const savedTasks = localStorage.getItem("tasks");
+    return savedTasks ? JSON.parse(savedTasks) : [];
+  });
+
+  useEffect(() => {
+    localStorage.setItem("tasks", JSON.stringify(tasks));
+  }, [tasks]);
+
+  const addTask = (title: string, priority: Task["priority"]) => {
+    const newTask: Task = {
+      id: Date.now().toString(),
+      title,
+      completed: false,
+      priority,
+    };
+    setTasks((prev) => [...prev, newTask]);
+  };
+
+  const toggleTask = (id: string) => {
+    setTasks((prev) =>
+      prev.map((task) =>
+        task.id === id ? { ...task, completed: !task.completed } : task
+      )
+    );
+  };
+
+  const deleteTask = (id: string) => {
+    setTasks((prev) => prev.filter((task) => task.id !== id));
+  };
+
+  return {
+    tasks,
+    addTask,
+    toggleTask,
+    deleteTask,
+  };
+}
